cachefilesd historical revision 0.8
Made cachefilesd ask the kernel whether cullable objects are in use and omit
them from the cull table if they are.

Made the size of cachefilesd's culling tables configurable.

Updated the manual pages.

Documented SELinux interaction.

Include SELinux policy for cachefilesd.

Fixed typo that was causing the howto.txt not to be installed.

Use /dev/cachefiles if it present in preference to /proc/fs/cachefiles.

Use poll rather than SIGURG on /dev/cachefilesd.