Provide an option to suppress cachefilesd culling
Provide an option to suppress cachefilesd culling.  This prevents the daemon
from using up resources scanning the cache.  The downside is that it never
attempts to shrink the cache to make space when the cache ends up exceeding
its limits due to other processes using up resources on the same filesystem as
the cache.

Signed-off-by: David Howells <dhowells@redhat.com>