wip: work in progress / temporary patch
Possibly not important; I believe I was working on different
loop structures and algorithms to try to make the culling process faster.

Was trying to optimize the read-insert-sort loop in general,
and I thought I'd try to see what happens when GCC encounters a function pointer
that is not likely to change again in the duration of the loop.

IIRC, the performance was terrible. :)

AFAIK, I was looking at writing different functions that were optimized for full
 and non-full insertions to see if that could be made faster, but I am not sure
I made much progress.